+++ tpFinal/microStation/src/main/java/com/microstation/controller/StationController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 
 import com.microstation.dto.StationDTO;
 import com.microstation.service.StationService;
@@ -24,13 +25,35 @@ public class StationController {
     @Autowired
     private StationService stationService;
 
+    // URL del servicio de validación de tokens
+    private static final String TOKEN_VALIDATION_URL = "http://localhost:8081/auth/validar";
+
+    /**
+     * validarToken
+     * Valida el token antes de realizar cualquier operación.
+     * @param token
+     */
+    private ResponseEntity<String> validarToken(String token) {
+        // Realizar una llamada al servicio de validación de tokens
+        // System.out.println("Validando token: " + token);
+        ResponseEntity<String> response = new RestTemplate().postForEntity(TOKEN_VALIDATION_URL, token, String.class);
+        // System.out.println("Respuesta: " + response);
+        return response;
+    }
+
+
     /**
      * Obtiene todas las estaciones.
      * @return ResponseEntity
      */
     @Operation(summary = "Obtiene todas las estaciones.", description = "Obtiene todos las estaciones")
     @GetMapping("")
-    public ResponseEntity<?> getAll(){
+    public ResponseEntity<?> getAll(@RequestHeader("Authorization") String token){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(stationService.findAll());
         }catch (Exception e){
@@ -44,7 +67,12 @@ public class StationController {
      */
     @Operation(summary = "Agrega una estacion.", description = "Agrega una estacion")
     @PostMapping("/alta")
-    public ResponseEntity<?> save(@RequestBody StationDTO entity){
+    public ResponseEntity<?> save(@RequestHeader("Authorization") String token,@RequestBody StationDTO entity){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{            
             return ResponseEntity.status(HttpStatus.OK).body(stationService.save(entity));
         }catch (Exception e){
@@ -59,7 +87,12 @@ public class StationController {
      */
     @Operation(summary = "Obtiene una estacion por su id.", description = "Obtiene un estacion por su stationId")
     @GetMapping("/buscar/{stationId}")
-    public ResponseEntity<?> getById(@PathVariable String stationId) {
+    public ResponseEntity<?> getById(@RequestHeader("Authorization") String token,@PathVariable String stationId) {
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(stationService.findById(stationId));
         }catch (Exception e){
@@ -74,7 +107,12 @@ public class StationController {
      */
     @Operation(summary = "Eliminia una estacion por su id.", description = "Elimina una estacion por su stationId")
     @DeleteMapping("/eliminar/{stationId}")
-    public ResponseEntity<?> delete(@PathVariable String stationId){
+    public ResponseEntity<?> delete(@RequestHeader("Authorization") String token,@PathVariable String stationId){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             stationService.delete(stationId);
             return ResponseEntity.status(HttpStatus.OK).body("Se elimino correctamente la estacion con stationId: " + stationId);
@@ -91,7 +129,13 @@ public class StationController {
      */
     @Operation(summary = "Actualiza los datos de una estacion por su id.", description = "Actualiza una estacion por su stationId")
     @PutMapping("/actualizar/{stationId}")
-    public ResponseEntity<?> update(@PathVariable String stationId, @RequestBody StationDTO entity){
+    public ResponseEntity<?> update(@RequestHeader("Authorization") String token,@PathVariable String stationId, @RequestBody StationDTO entity){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
+
         try{
             stationService.update(stationId, entity);
             return ResponseEntity.status(HttpStatus.OK).body("Se actualizaron correctamente los datos de la estacion con stationId: " + stationId);
@@ -108,7 +152,12 @@ public class StationController {
      */
     @Operation(summary = "Verifica si las coordenadas son validas.", description = "Verifica que las coordenadas proveidas coinciden con als coordenadas de una estacion")
     @GetMapping("/verificar/latitud/{latitud}/longitud/{longitud}")
-    public ResponseEntity<?> verify(@PathVariable String latitud, @PathVariable String longitud){
+    public ResponseEntity<?> verify(@RequestHeader("Authorization") String token,@PathVariable String latitud, @PathVariable String longitud){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(stationService.findByLatitudAndLongitud(latitud, longitud));
         }catch (IllegalArgumentException e){
