+++ tpFinal/microUserAccount/src/main/java/com/microuseraccount/controller/AccountController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 
 import com.microuseraccount.dto.AccountDTO;
 import com.microuseraccount.service.AccountService;
@@ -23,6 +24,23 @@ public class AccountController {
     @Autowired
     private AccountService accountService;
 
+    // URL del servicio de validación de tokens
+    private static final String TOKEN_VALIDATION_URL = "http://localhost:8081/auth/validar";
+
+    /**
+     * validarToken
+     * Valida el token antes de realizar cualquier operación.
+     * @param token
+     */
+    private ResponseEntity<String> validarToken(String token) {
+        // Realizar una llamada al servicio de validación de tokens
+        // System.out.println("Validando token: " + token);
+        ResponseEntity<String> response = new RestTemplate().postForEntity(TOKEN_VALIDATION_URL, token, String.class);
+        // System.out.println("Respuesta: " + response);
+        return response;
+    }
+
+
     /**
      * getAll
      * Obtiene todas las cuentas.
@@ -30,7 +48,12 @@ public class AccountController {
      */
     @Operation(summary = "Obtiene todos las cuentas.", description = "Obtiene todas las cuentas")
     @GetMapping("")
-    public ResponseEntity<?> getAll(){
+    public ResponseEntity<?> getAll(@RequestHeader("Authorization") String token){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(accountService.findAll());
         }catch (Exception e){
@@ -46,7 +69,12 @@ public class AccountController {
      */
     @Operation(summary = "Obtiene todas las cuentas de un usuario.", description = "Obtiene todas las cuentas de un usuario")
     @GetMapping("/usuario/{userId}")
-    public ResponseEntity<?> getCuentasByUserId(@PathVariable long userId){
+    public ResponseEntity<?> getCuentasByUserId(@RequestHeader("Authorization") String token, @PathVariable long userId){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(accountService.getCuentasByUserId(userId));
         }catch (Exception e){
@@ -62,7 +90,12 @@ public class AccountController {
      */
     @Operation(summary = "Crea una nueva cuenta.", description = "Crea una cuenta")
     @PostMapping("/alta")
-    public ResponseEntity<?> save(@RequestBody AccountDTO entity){
+    public ResponseEntity<?> save(@RequestHeader("Authorization") String token, @RequestBody AccountDTO entity){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{            
             return ResponseEntity.status(HttpStatus.OK).body(accountService.save(entity));
         }catch (Exception e){
@@ -78,7 +111,12 @@ public class AccountController {
      */
     @Operation(summary = "Obtiene una cuenta por su id.", description = "Obtiene una cuenta por su accountId")
     @GetMapping("/buscar/{accountId}")
-    public ResponseEntity<?> getById(@PathVariable long accountId) {
+    public ResponseEntity<?> getById(@RequestHeader("Authorization") String token, @PathVariable long accountId) {
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(accountService.findById(accountId));
         }catch (Exception e){
@@ -94,7 +132,12 @@ public class AccountController {
      */
     @Operation(summary = "Elimina una cuenta por su id.", description = "Elimina una cuenta por su accountId")    
     @DeleteMapping("/eliminar/{accountId}")
-    public ResponseEntity<?> delete(@PathVariable long accountId){
+    public ResponseEntity<?> delete(@RequestHeader("Authorization") String token, @PathVariable long accountId){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             accountService.delete(accountId);
             return ResponseEntity.status(HttpStatus.OK).body("Se elimino correctamente la cuenta con accountId: " + accountId);
@@ -111,7 +154,12 @@ public class AccountController {
      */
     @Operation(summary = "Desactiva una cuenta por su id.", description = "Desactiva una cuenta por su accountId")
     @PutMapping("/suspender/{accountId}")
-    public ResponseEntity<?> suspend(@PathVariable long accountId){
+    public ResponseEntity<?> suspend(@RequestHeader("Authorization") String token, @PathVariable long accountId){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             accountService.suspendAccount(accountId);
             return ResponseEntity.status(HttpStatus.OK).body("Se suspendio correctamente la cuenta con accountId: " + accountId);
@@ -128,7 +176,12 @@ public class AccountController {
      */
     @Operation(summary = "Activa una cuenta por su id.", description = "Activa una cuenta por su accountId")
     @PutMapping("/activar/{accountId}")
-    public ResponseEntity<?> activate(@PathVariable long accountId){
+    public ResponseEntity<?> activate(@RequestHeader("Authorization") String token, @PathVariable long accountId){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             accountService.activateAccount(accountId);
             return ResponseEntity.status(HttpStatus.OK).body("Se activo correctamente la cuenta con accountId: " + accountId);
@@ -146,7 +199,12 @@ public class AccountController {
      */
     @Operation(summary = "Actualiza una cuenta por su id.", description = "Actualiza una cuenta por su accountId")
     @PutMapping("/actualizar/{accountId}")
-    public ResponseEntity<?> update(@PathVariable long accountId, @RequestBody AccountDTO entity){
+    public ResponseEntity<?> update(@RequestHeader("Authorization") String token, @PathVariable long accountId, @RequestBody AccountDTO entity){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             accountService.update(accountId, entity);
             return ResponseEntity.status(HttpStatus.OK).body("Se actualizo correctamente la cuenta con accountId: " + accountId);
@@ -164,7 +222,12 @@ public class AccountController {
      */
     @Operation(summary = "Vincula una cuenta a un usuario.", description = "Vincula un usuario a una cuenta")
     @PutMapping("/vincular/usuario/{userId}/cuenta/{accountId}")
-    public ResponseEntity<?> asociarUsuario(@PathVariable long userId, @PathVariable long accountId){
+    public ResponseEntity<?> asociarUsuario(@RequestHeader("Authorization") String token, @PathVariable long userId, @PathVariable long accountId){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             accountService.asociarUsuario(userId, accountId);
             return ResponseEntity.status(HttpStatus.OK).body("Se asocio correctamente el usuario con userId: " + userId + " a la cuenta con accountId: " + accountId);
@@ -182,7 +245,12 @@ public class AccountController {
      */
     @Operation(summary = "Desvincula una cuenta de un usuario.", description = "Desvincula un usuario de una cuenta")
     @DeleteMapping("/desvincular/usuario/{userId}/cuenta/{accountId}")
-    public ResponseEntity<?> desvincularUsuario(@PathVariable long userId, @PathVariable long accountId){
+    public ResponseEntity<?> desvincularUsuario(@RequestHeader("Authorization") String token, @PathVariable long userId, @PathVariable long accountId){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             accountService.desvincularUsuario(userId, accountId);
             return ResponseEntity.status(HttpStatus.OK).body("Se desvinculo correctamente el usuario con userId: " + userId + " de la cuenta con accountId: " + accountId);
@@ -199,7 +267,12 @@ public class AccountController {
      */
     @Operation(summary = "Obtiene el saldo de una cuenta.", description = "Obtiene el saldo de una cuenta")
     @GetMapping("/saldo/obtener/{accountId}")
-    public ResponseEntity<?> getSaldo(@PathVariable long accountId) {
+    public ResponseEntity<?> getSaldo(@RequestHeader("Authorization") String token, @PathVariable long accountId) {
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(accountService.getSaldo(accountId));
         }catch (Exception e){
@@ -215,7 +288,17 @@ public class AccountController {
      */
     @Operation(summary = "Actualiza el saldo de una cuenta.", description = "Actualiza el saldo de una cuenta")
     @PutMapping("/saldo/actualizar/{accountId}")
-    public void updateSaldo(@PathVariable long accountId, @RequestBody Double saldo) {
-        accountService.updateSaldo(saldo, accountId);
+    public ResponseEntity<?> updateSaldo(@RequestHeader("Authorization") String token, @PathVariable long accountId, @RequestBody Double saldo) {
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
+        try{
+            accountService.updateSaldo(saldo, accountId);
+            return ResponseEntity.status(HttpStatus.OK).body("Saldo Actualizado!");
+        }catch(Exception e){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{\"error\":\"Error. Intente nuevamente.\"\n\"error\":\"" + e.getMessage()+"\"}");
+        }
     }
 }
